Exponentially-weighted moving average/std/var.

Exponentially-weighted moving average/std/var.

Ewm_std

Ewm_var

Expr_ewm_mean(
  com = NULL,
  span = NULL,
  half_life = NULL,
  alpha = NULL,
  adjust = TRUE,
  min_periods = 1L
)

Expr_ewm_std(
  com = NULL,
  span = NULL,
  half_life = NULL,
  alpha = NULL,
  adjust = TRUE,
  bias = FALSE,
  min_periods = 1L
)

Expr_ewm_var(
  com = NULL,
  span = NULL,
  half_life = NULL,
  alpha = NULL,
  adjust = TRUE,
  bias = FALSE,
  min_periods = 1L
)

Format

Method

Arguments

com

Specify decay in terms of center of mass, \(\gamma\), with \( \alpha = \frac{1}{1 + \gamma} \; \forall \; \gamma \geq 0 \)

span

Specify decay in terms of span, \(\theta\), with \(\alpha = \frac{2}{\theta + 1} \; \forall \; \theta \geq 1 \)

half_life

Specify decay in terms of half-life, :math:\lambda, with \( \alpha = 1 - \exp \left\{ \frac{ -\ln(2) }{ \lambda } \right\} \) \( \forall \; \lambda > 0\)

alpha

Specify smoothing factor alpha directly, \(0 < \alpha \leq 1\).

adjust

Divide by decaying adjustment factor in beginning periods to account for imbalance in relative weightings

When adjust=True the EW function is calculated using weights \(w_i = (1 - \alpha)^i \)

When adjust=False the EW function is calculated recursively by \( y_0 = x_0 \\ y_t = (1 - \alpha)y_{t - 1} + \alpha x_t \)

min_periods

Minimum number of observations in window required to have a value (otherwise result is null).

bias

When bias=FALSE`, apply a correction to make the estimate statistically unbiased.

Value

Expr

Examples

pl$DataFrame(a = 1:3)$select(pl$col("a")$ewm_mean(com=1))
#> polars DataFrame: shape: (3, 1)
#> ┌──────────┐
#> │ a        │
#> │ ---      │
#> │ f64      │
#> ╞══════════╡
#> │ 1.0      │
#> ├╌╌╌╌╌╌╌╌╌╌┤
#> │ 1.666667 │
#> ├╌╌╌╌╌╌╌╌╌╌┤
#> │ 2.428571 │
#> └──────────┘

pl$DataFrame(a = 1:3)$select(pl$col("a")$ewm_std(com=1))
#> polars DataFrame: shape: (3, 1)
#> ┌──────────┐
#> │ a        │
#> │ ---      │
#> │ f64      │
#> ╞══════════╡
#> │ 0.0      │
#> ├╌╌╌╌╌╌╌╌╌╌┤
#> │ 0.707107 │
#> ├╌╌╌╌╌╌╌╌╌╌┤
#> │ 0.963624 │
#> └──────────┘
pl$DataFrame(a = 1:3)$select(pl$col("a")$ewm_std(com=1))
#> polars DataFrame: shape: (3, 1)
#> ┌──────────┐
#> │ a        │
#> │ ---      │
#> │ f64      │
#> ╞══════════╡
#> │ 0.0      │
#> ├╌╌╌╌╌╌╌╌╌╌┤
#> │ 0.707107 │
#> ├╌╌╌╌╌╌╌╌╌╌┤
#> │ 0.963624 │
#> └──────────┘