Inner workings of the Series-class

The Series-class is simply two environments of respectively the public and private methods/function calls to the rpolars rust side. The instanciated Series-object is an externalptr to a lowlevel rust polars Series object. The pointer address is the only statefullness of the Series object on the R side. Any other state resides on the rust side. The S3 method .DollarNames.Series exposes all public $foobar()-methods which are callable onto the object. Most methods return another Series-class instance or similar which allows for method chaining. This class system in lack of a better name could be called "environment classes" and is the same class system extendr provides, except here there is both a public and private set of methods. For implementation reasons, the private methods are external and must be called from rpolars:::.pr.$Series$methodname(), also all private methods must take any self as an argument, thus they are pure functions. Having the private methods as pure functions solved/simplified self-referential complications.

Details

Check out the source code in R/Series_frame.R how public methods are derived from private methods. Check out extendr-wrappers.R to see the extendr-auto-generated methods. These are moved to .pr and converted into pure external functions in after-wrappers.R. In zzz.R (named zzz to be last file sourced) the extendr-methods are removed and replaced by any function prefixed Series_.

Examples

#see all exported methods
ls(rpolars:::Series)
#>  [1] "abs"           "add"           "alias"         "all"          
#>  [5] "any"           "append"        "apply"         "arg_max"      
#>  [9] "arg_min"       "ceil"          "chunk_lengths" "clone"        
#> [13] "compare"       "cumsum"        "div"           "dtype"        
#> [17] "flags"         "floor"         "is_numeric"    "is_unique"    
#> [21] "len"           "max"           "min"           "mul"          
#> [25] "name"          "print"         "rem"           "rename"       
#> [29] "rep"           "series_equal"  "shape"         "sort"         
#> [33] "std"           "sub"           "sum"           "to_frame"     
#> [37] "to_r"          "to_r_list"     "to_r_vector"   "value_counts" 
#> [41] "var"          

#see all private methods (not intended for regular use)
ls(rpolars:::.pr$Series)
#>  [1] "abs"                    "add"                    "alias"                 
#>  [4] "all"                    "any"                    "append_mut"            
#>  [7] "apply"                  "arg_max"                "arg_min"               
#> [10] "ceil"                   "chunk_lengths"          "clone"                 
#> [13] "compare"                "cumsum"                 "div"                   
#> [16] "dtype"                  "floor"                  "is_sorted_flag"        
#> [19] "is_sorted_reverse_flag" "is_unique"              "len"                   
#> [22] "max"                    "min"                    "mul"                   
#> [25] "name"                   "new"                    "print"                 
#> [28] "rem"                    "rename_mut"             "rep"                   
#> [31] "series_equal"           "shape"                  "sort_mut"              
#> [34] "sub"                    "sum"                    "to_frame"              
#> [37] "to_r"                   "value_counts"          

#make an object
s = pl$Series(1:3)

#use a public method/property
s$shape
#> [1] 3 1

#use a private method (mutable append not allowed in public api)
s_copy = s
.pr$Series$append_mut(s, pl$Series(5:1))
#> $ok
#> NULL
#> 
#> $err
#> NULL
#> 
#> attr(,"class")
#> [1] "Result"
identical(s_copy$to_r(), s$to_r()) # s_copy was modified when s was modified
#> [1] TRUE