generic setter method

set value of properties of DataFrames

# S3 method for DataFrame
$(self, name) <- value

Value

value

Details

settable rpolars object properties may appear to be R objects, but they are not. See [method_name] example

Examples

#For internal use
#is only activated for following methods of DataFrame
ls(rpolars:::DataFrame.property_setters)
#> [1] "columns"

#specific use case for one object property 'columns' (names)
df = pl$DataFrame(iris)

#get values
df$columns
#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"     

#set + get values
df$columns = letters[1:5] #<- is fine too
df$columns
#> [1] "a" "b" "c" "d" "e"

# Rstudio is not using the standard R code completion tool
# and it will backtick any special characters. It is possible
# to completely customize the R / Rstudio code completion except
# it will trigger Rstudio to backtick any completion! Also R does
# not support package isolated customization.

#Concrete example if tabbing on 'df$' the raw R suggestion is df$columns<-
#however Rstudio backticks it into df$`columns<-`
#to make life simple, this is valid rpolars syntax also, and can be used in fast scripting
df$`columns<-` = letters[5:1]

#for stable code prefer e.g.  df$columns = letters[5:1]

#to see inside code of a property use the [[]] syntax instead
df[["columns"]] # to see property code, .pr is the internal rpolars api into rust polars
#> function() {
#>   .pr$DataFrame$columns(self)
#> }
#> <environment: 0x559216f72ff8>
#> attr(,"class")
#> [1] "setter"   "property" "function"
rpolars:::DataFrame.property_setters$columns #and even more obscure to see setter code
#> function(self, names) unwrap(.pr$DataFrame$set_column_names_mut(self,names))
#> <environment: namespace:rpolars>