Start Expression with a column

Return an expression representing a column in a DataFrame.

Arguments

name

a single column by a string

all columns by using a wildcard "*"

multiple columns as vector of strings

column by regular expression if the regex starts with ^ and ends with $ e.g. pl$DataFrame(iris)$select(pl$col(c("^Sepal.*$")))

a single DataType or an R list of DataTypes, select any column of any such DataType

Series of utf8 strings abiding to above options

Value

Column Exprression

Examples

df = pl$DataFrame(list(foo=1, bar=2L,foobar="3"))

#a single column by a string
df$select(pl$col("foo"))
#> polars DataFrame: shape: (1, 1)
#> ┌─────┐
#> │ foo │
#> │ --- │
#> │ f64 │
#> ╞═════╡
#> │ 1.0 │
#> └─────┘

#all columns by wildcard
df$select(pl$col("*"))
#> polars DataFrame: shape: (1, 3)
#> ┌─────┬─────┬────────┐
#> │ foo ┆ bar ┆ foobar │
#> │ --- ┆ --- ┆ ---    │
#> │ f64 ┆ i32 ┆ str    │
#> ╞═════╪═════╪════════╡
#> │ 1.0 ┆ 2   ┆ 3      │
#> └─────┴─────┴────────┘
df$select(pl$all())
#> polars DataFrame: shape: (1, 3)
#> ┌─────┬─────┬────────┐
#> │ foo ┆ bar ┆ foobar │
#> │ --- ┆ --- ┆ ---    │
#> │ f64 ┆ i32 ┆ str    │
#> ╞═════╪═════╪════════╡
#> │ 1.0 ┆ 2   ┆ 3      │
#> └─────┴─────┴────────┘

#multiple columns as vector of strings
df$select(pl$col(c("foo","bar")))
#> polars DataFrame: shape: (1, 2)
#> ┌─────┬─────┐
#> │ foo ┆ bar │
#> │ --- ┆ --- │
#> │ f64 ┆ i32 │
#> ╞═════╪═════╡
#> │ 1.0 ┆ 2   │
#> └─────┴─────┘

#column by regular expression if the regex starts with `^` and ends with `$`
df$select(pl$col("^foo.*$"))
#> polars DataFrame: shape: (1, 2)
#> ┌─────┬────────┐
#> │ foo ┆ foobar │
#> │ --- ┆ ---    │
#> │ f64 ┆ str    │
#> ╞═════╪════════╡
#> │ 1.0 ┆ 3      │
#> └─────┴────────┘

#a single DataType
df$select(pl$col(pl$dtypes$Float64))
#> polars DataFrame: shape: (1, 1)
#> ┌─────┐
#> │ foo │
#> │ --- │
#> │ f64 │
#> ╞═════╡
#> │ 1.0 │
#> └─────┘

# ... or an R list of DataTypes, select any column of any such DataType
df$select(pl$col(list(pl$dtypes$Float64, pl$dtypes$Utf8)))
#> polars DataFrame: shape: (1, 2)
#> ┌─────┬────────┐
#> │ foo ┆ foobar │
#> │ --- ┆ ---    │
#> │ f64 ┆ str    │
#> ╞═════╪════════╡
#> │ 1.0 ┆ 3      │
#> └─────┴────────┘

# from Series of names
df$select(pl$col(pl$Series(c("bar","foobar"))))
#> polars DataFrame: shape: (1, 2)
#> ┌─────┬────────┐
#> │ bar ┆ foobar │
#> │ --- ┆ ---    │
#> │ i32 ┆ str    │
#> ╞═════╪════════╡
#> │ 2   ┆ 3      │
#> └─────┴────────┘