polars-API: private calls to rust-polars

.pr Original extendr bindings converted into pure functions

Usage

.pr

Format

An object of class environment of length 11.

Examples

#.pr$DataFrame$print() is an external function where self is passed as arg
rpolars:::.pr$DataFrame$print(self = pl$DataFrame(iris))
#> shape: (150, 5)
#> ┌──────────────┬─────────────┬──────────────┬─────────────┬───────────┐
#> │ Sepal.Length ┆ Sepal.Width ┆ Petal.Length ┆ Petal.Width ┆ Species   │
#> │ ---          ┆ ---         ┆ ---          ┆ ---         ┆ ---       │
#> │ f64          ┆ f64         ┆ f64          ┆ f64         ┆ cat       │
#> ╞══════════════╪═════════════╪══════════════╪═════════════╪═══════════╡
#> │ 5.1          ┆ 3.5         ┆ 1.4          ┆ 0.2         ┆ setosa    │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 4.9          ┆ 3.0         ┆ 1.4          ┆ 0.2         ┆ setosa    │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 4.7          ┆ 3.2         ┆ 1.3          ┆ 0.2         ┆ setosa    │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 4.6          ┆ 3.1         ┆ 1.5          ┆ 0.2         ┆ setosa    │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ ...          ┆ ...         ┆ ...          ┆ ...         ┆ ...       │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 6.3          ┆ 2.5         ┆ 5.0          ┆ 1.9         ┆ virginica │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 6.5          ┆ 3.0         ┆ 5.2          ┆ 2.0         ┆ virginica │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 6.2          ┆ 3.4         ┆ 5.4          ┆ 2.3         ┆ virginica │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 5.9          ┆ 3.0         ┆ 5.1          ┆ 1.8         ┆ virginica │
#> └──────────────┴─────────────┴──────────────┴─────────────┴───────────┘
#> polars DataFrame: shape: (150, 5)
#> ┌──────────────┬─────────────┬──────────────┬─────────────┬───────────┐
#> │ Sepal.Length ┆ Sepal.Width ┆ Petal.Length ┆ Petal.Width ┆ Species   │
#> │ ---          ┆ ---         ┆ ---          ┆ ---         ┆ ---       │
#> │ f64          ┆ f64         ┆ f64          ┆ f64         ┆ cat       │
#> ╞══════════════╪═════════════╪══════════════╪═════════════╪═══════════╡
#> │ 5.1          ┆ 3.5         ┆ 1.4          ┆ 0.2         ┆ setosa    │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 4.9          ┆ 3.0         ┆ 1.4          ┆ 0.2         ┆ setosa    │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 4.7          ┆ 3.2         ┆ 1.3          ┆ 0.2         ┆ setosa    │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 4.6          ┆ 3.1         ┆ 1.5          ┆ 0.2         ┆ setosa    │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ ...          ┆ ...         ┆ ...          ┆ ...         ┆ ...       │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 6.3          ┆ 2.5         ┆ 5.0          ┆ 1.9         ┆ virginica │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 6.5          ┆ 3.0         ┆ 5.2          ┆ 2.0         ┆ virginica │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 6.2          ┆ 3.4         ┆ 5.4          ┆ 2.3         ┆ virginica │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┤
#> │ 5.9          ┆ 3.0         ┆ 5.1          ┆ 1.8         ┆ virginica │
#> └──────────────┴─────────────┴──────────────┴─────────────┴───────────┘

rpolars:::print_env(.pr,".pr the collection of private method calls to rust-polars")
#> 
#> 
#>     .pr the collection of private method calls to rust-polars ( environment ):
#> 
#>        DataFrame ( environment ):
#>           [ by_agg ; function ]
#>           [ clone_see_me_macro ; function ]
#>           [ columns ; function ]
#>           [ dtypes ; function ]
#>           [ get_column ; function ]
#>           [ get_columns ; function ]
#>           [ lazy ; function ]
#>           [ new ; function ]
#>           [ new_with_capacity ; function ]
#>           [ print ; function ]
#>           [ schema ; function ]
#>           [ select ; function ]
#>           [ select_at_idx ; function ]
#>           [ set_column_from_robj ; function ]
#>           [ set_column_from_series ; function ]
#>           [ set_column_names_mut ; function ]
#>           [ shape ; function ]
#>           [ to_list ; function ]
#>           [ to_list_tag_structs ; function ]
#>           [ to_struct ; function ]
#>           [ unnest ; function ]
#> 
#> 
#>        DataType ( environment ):
#>           [ eq ; function ]
#>           [ get_all_simple_type_names ; function ]
#>           [ ne ; function ]
#>           [ new ; function ]
#>           [ new_datetime ; function ]
#>           [ new_duration ; function ]
#>           [ new_list ; function ]
#>           [ new_object ; function ]
#>           [ new_struct ; function ]
#>           [ print ; function ]
#> 
#> 
#>        DataTypeVector ( environment ):
#>           [ from_rlist ; function ]
#>           [ new ; function ]
#>           [ print ; function ]
#>           [ push ; function ]
#> 
#> 
#>        Expr ( environment ):
#>           [ abs ; function ]
#>           [ add ; function ]
#>           [ agg_groups ; function ]
#>           [ alias ; function ]
#>           [ all ; function ]
#>           [ and ; function ]
#>           [ any ; function ]
#>           [ append ; function ]
#>           [ arccos ; function ]
#>           [ arccosh ; function ]
#>           [ arcsin ; function ]
#>           [ arcsinh ; function ]
#>           [ arctan ; function ]
#>           [ arctanh ; function ]
#>           [ arg_max ; function ]
#>           [ arg_min ; function ]
#>           [ arg_sort ; function ]
#>           [ arg_unique ; function ]
#>           [ backward_fill ; function ]
#>           [ cast ; function ]
#>           [ ceil ; function ]
#>           [ clip ; function ]
#>           [ clip_max ; function ]
#>           [ clip_min ; function ]
#>           [ col ; function ]
#>           [ cols ; function ]
#>           [ cos ; function ]
#>           [ cosh ; function ]
#>           [ count ; function ]
#>           [ cumcount ; function ]
#>           [ cummax ; function ]
#>           [ cummin ; function ]
#>           [ cumprod ; function ]
#>           [ cumsum ; function ]
#>           [ cumulative_eval ; function ]
#>           [ diff ; function ]
#>           [ div ; function ]
#>           [ dot ; function ]
#>           [ drop_nans ; function ]
#>           [ drop_nulls ; function ]
#>           [ dtype_cols ; function ]
#>           [ entropy ; function ]
#>           [ eq ; function ]
#>           [ ewm_mean ; function ]
#>           [ ewm_std ; function ]
#>           [ ewm_var ; function ]
#>           [ exclude ; function ]
#>           [ exclude_dtype ; function ]
#>           [ exp ; function ]
#>           [ explode ; function ]
#>           [ extend_constant ; function ]
#>           [ extend_expr ; function ]
#>           [ fill_nan ; function ]
#>           [ fill_null ; function ]
#>           [ fill_null_with_strategy ; function ]
#>           [ filter ; function ]
#>           [ first ; function ]
#>           [ flatten ; function ]
#>           [ floor ; function ]
#>           [ forward_fill ; function ]
#>           [ gt ; function ]
#>           [ gt_eq ; function ]
#>           [ hash ; function ]
#>           [ head ; function ]
#>           [ interpolate ; function ]
#>           [ is_duplicated ; function ]
#>           [ is_finite ; function ]
#>           [ is_first ; function ]
#>           [ is_in ; function ]
#>           [ is_infinite ; function ]
#>           [ is_nan ; function ]
#>           [ is_not ; function ]
#>           [ is_not_nan ; function ]
#>           [ is_not_null ; function ]
#>           [ is_null ; function ]
#>           [ is_unique ; function ]
#>           [ keep_name ; function ]
#>           [ kurtosis ; function ]
#>           [ last ; function ]
#>           [ len ; function ]
#>           [ list ; function ]
#>           [ lit ; function ]
#>           [ log ; function ]
#>           [ log10 ; function ]
#>           [ lower_bound ; function ]
#>           [ lt ; function ]
#>           [ lt_eq ; function ]
#>           [ map ; function ]
#>           [ map_alias ; function ]
#>           [ max ; function ]
#>           [ mean ; function ]
#>           [ median ; function ]
#>           [ min ; function ]
#>           [ mode ; function ]
#>           [ mul ; function ]
#>           [ n_unique ; function ]
#>           [ nan_max ; function ]
#>           [ nan_min ; function ]
#>           [ neq ; function ]
#>           [ null_count ; function ]
#>           [ or ; function ]
#>           [ over ; function ]
#>           [ pct_change ; function ]
#>           [ pow ; function ]
#>           [ prefix ; function ]
#>           [ print ; function ]
#>           [ product ; function ]
#>           [ quantile ; function ]
#>           [ rank ; function ]
#>           [ rechunk ; function ]
#>           [ reinterpret ; function ]
#>           [ rep ; function ]
#>           [ repeat_by ; function ]
#>           [ reshape ; function ]
#>           [ reverse ; function ]
#>           [ rolling_max ; function ]
#>           [ rolling_mean ; function ]
#>           [ rolling_median ; function ]
#>           [ rolling_min ; function ]
#>           [ rolling_quantile ; function ]
#>           [ rolling_skew ; function ]
#>           [ rolling_std ; function ]
#>           [ rolling_sum ; function ]
#>           [ rolling_var ; function ]
#>           [ round ; function ]
#>           [ sample_frac ; function ]
#>           [ sample_n ; function ]
#>           [ search_sorted ; function ]
#>           [ shift ; function ]
#>           [ shift_and_fill ; function ]
#>           [ shrink_dtype ; function ]
#>           [ shuffle ; function ]
#>           [ sign ; function ]
#>           [ sin ; function ]
#>           [ sinh ; function ]
#>           [ skew ; function ]
#>           [ slice ; function ]
#>           [ sort ; function ]
#>           [ sort_by ; function ]
#>           [ std ; function ]
#>           [ sub ; function ]
#>           [ suffix ; function ]
#>           [ sum ; function ]
#>           [ tail ; function ]
#>           [ take ; function ]
#>           [ take_every ; function ]
#>           [ tan ; function ]
#>           [ tanh ; function ]
#>           [ to_physical ; function ]
#>           [ top_k ; function ]
#>           [ unique ; function ]
#>           [ unique_counts ; function ]
#>           [ unique_stable ; function ]
#>           [ upper_bound ; function ]
#>           [ value_counts ; function ]
#>           [ var ; function ]
#>           [ xor ; function ]
#> 
#>        [ GroupBy ; NULL ]
#> 
#>        LazyFrame ( environment ):
#>           [ collect ; function ]
#>           [ describe_optimized_plan ; function ]
#>           [ describe_plan ; function ]
#>           [ filter ; function ]
#>           [ groupby ; function ]
#>           [ join ; function ]
#>           [ limit ; function ]
#>           [ print ; function ]
#>           [ select ; function ]
#>           [ with_column ; function ]
#>           [ with_columns ; function ]
#> 
#> 
#>        LazyGroupBy ( environment ):
#>           [ agg ; function ]
#>           [ head ; function ]
#>           [ print ; function ]
#>           [ tail ; function ]
#> 
#> 
#>        ProtoExprArray ( environment ):
#>           [ new ; function ]
#>           [ print ; function ]
#>           [ push_back_rexpr ; function ]
#>           [ push_back_str ; function ]
#> 
#> 
#>        RNullValues ( environment ):
#>           [ new_all_columns ; function ]
#>           [ new_columns ; function ]
#>           [ new_named ; function ]
#> 
#> 
#>        Series ( environment ):
#>           [ abs ; function ]
#>           [ add ; function ]
#>           [ alias ; function ]
#>           [ all ; function ]
#>           [ any ; function ]
#>           [ append_mut ; function ]
#>           [ apply ; function ]
#>           [ arg_max ; function ]
#>           [ arg_min ; function ]
#>           [ ceil ; function ]
#>           [ chunk_lengths ; function ]
#>           [ clone ; function ]
#>           [ compare ; function ]
#>           [ cumsum ; function ]
#>           [ div ; function ]
#>           [ dtype ; function ]
#>           [ floor ; function ]
#>           [ is_sorted_flag ; function ]
#>           [ is_sorted_reverse_flag ; function ]
#>           [ is_unique ; function ]
#>           [ len ; function ]
#>           [ max ; function ]
#>           [ min ; function ]
#>           [ mul ; function ]
#>           [ name ; function ]
#>           [ new ; function ]
#>           [ panic ; function ]
#>           [ print ; function ]
#>           [ rem ; function ]
#>           [ rename_mut ; function ]
#>           [ rep ; function ]
#>           [ series_equal ; function ]
#>           [ set_sorted_mut ; function ]
#>           [ shape ; function ]
#>           [ sleep ; function ]
#>           [ sort_mut ; function ]
#>           [ sub ; function ]
#>           [ sum ; function ]
#>           [ to_frame ; function ]
#>           [ to_r ; function ]
#>           [ value_counts ; function ]
#> 
#> 
#>        VecDataFrame ( environment ):
#>           [ new ; function ]
#>           [ print ; function ]
#>           [ push ; function ]
#>           [ with_capacity ; function ]
#>