The complete rpolars public API.

pl-object is a environment of all public functions and class constructors. Public functions are not exported as a normal package as it would be huge namespace collision with base:: and other functions. All object-methods are accessed with object$method() via the new class functions.

Having all functions in an namespace is similar to the rust- and python- polars api.

pl

Format

An object of class environment of length 34.

Details

If someone do not particularly like the letter combination pl, they are free to bind the environment to another variable name as simon_says = pl or even do attach(pl)

Examples

#how to use polars via `pl`
pl$col("colname")$sum() / pl$lit(42L)  #expression ~ chain-method / literal-expression
#> polars Expr: [(col("colname").sum()) / (42i32)]

#pl inventory
rpolars:::print_env(pl,"rpolars public functions")
#> 
#> 
#>     rpolars public functions ( environment ):
#>        [ Boolean ; DataType ]
#>        [ Categorical ; DataType ]
#>        [ DataFrame ; function ]
#>        [ Float32 ; DataType ]
#>        [ Float64 ; DataType ]
#>        [ Int32 ; DataType ]
#>        [ Int64 ; DataType ]
#>        [ Series ; function ]
#>        [ UInt32 ; DataType ]
#>        [ UInt64 ; DataType ]
#>        [ Utf8 ; DataType ]
#>        [ all ; function ]
#>        [ coalesce ; function ]
#>        [ col ; function ]
#>        [ concat ; function ]
#>        [ csv_reader ; function ]
#> 
#>        dtypes ( list ):
#>           [ Boolean ; DataType ]
#>           [ Categorical ; DataType ]
#>           [ Float32 ; DataType ]
#>           [ Float64 ; DataType ]
#>           [ Int32 ; DataType ]
#>           [ Int64 ; DataType ]
#>           [ UInt32 ; DataType ]
#>           [ UInt64 ; DataType ]
#>           [ Utf8 ; DataType ]
#> 
#>        [ expr_to_r ; function ]
#>        [ get_rpolars_opt_requirements ; function ]
#>        [ get_rpolars_options ; function ]
#>        [ lazy_csv_reader ; function ]
#>        [ lit ; function ]
#>        [ max ; function ]
#>        [ mem_address ; function ]
#>        [ min ; function ]
#> 
#>        numeric_dtypes ( list ):
#>           [ Float32 ; DataType ]
#>           [ Float64 ; DataType ]
#>           [ Int32 ; DataType ]
#>           [ Int64 ; DataType ]
#> 
#>        [ read_csv ; function ]
#>        [ rep ; function ]
#>        [ reset_rpolars_options ; function ]
#>        [ select ; function ]
#>        [ set_rpolars_options ; function ]
#>        [ std ; function ]
#>        [ sum ; function ]
#>        [ var ; function ]

#all accessible classes and their public methods
rpolars:::print_env(rpolars:::pl_pub_class_env,"rpolars public class methods, access via object$method()")
#> 
#> 
#>     rpolars public class methods, access via object$method() ( environment ):
#> 
#>        DataFrame ( environment ):
#>           [ as_data_frame ; function ]
#>           [ clone ; function ]
#>           [ columns ; setter property function ]
#>           [ dtypes ; property function ]
#>           [ filter ; function ]
#>           [ get_column ; function ]
#>           [ get_columns ; function ]
#>           [ groupby ; function ]
#>           [ height ; property function ]
#>           [ join ; function ]
#>           [ lazy ; function ]
#>           [ limit ; function ]
#>           [ print ; function ]
#>           [ schema ; property function ]
#>           [ select ; function ]
#>           [ shape ; property function ]
#>           [ to_data_frame ; function ]
#>           [ to_list ; function ]
#>           [ to_series ; function ]
#>           [ to_struct ; function ]
#>           [ unnest ; function ]
#>           [ width ; property function ]
#>           [ with_column ; function ]
#>           [ with_columns ; function ]
#> 
#> 
#>        DataType ( environment ):
#>           [ eq ; function ]
#>           [ get_all_type_names ; function ]
#>           [ ne ; function ]
#>           [ new ; function ]
#>           [ print ; function ]
#> 
#> 
#>        Expr ( environment ):
#>           [ abs ; function ]
#>           [ add ; function ]
#>           [ agg_groups ; function ]
#>           [ alias ; function ]
#>           [ all ; function ]
#>           [ and ; function ]
#>           [ any ; function ]
#>           [ append ; function ]
#>           [ apply ; function ]
#>           [ arccos ; function ]
#>           [ arccosh ; function ]
#>           [ arcsin ; function ]
#>           [ arcsinh ; function ]
#>           [ arctan ; function ]
#>           [ arctanh ; function ]
#>           [ arg_max ; function ]
#>           [ arg_min ; function ]
#>           [ arg_sort ; function ]
#>           [ arg_unique ; function ]
#>           [ argsort ; function ]
#>           [ backward_fill ; function ]
#>           [ cast ; function ]
#>           [ ceil ; function ]
#>           [ clip ; function ]
#>           [ clip_max ; function ]
#>           [ clip_min ; function ]
#>           [ cos ; function ]
#>           [ cosh ; function ]
#>           [ count ; function ]
#>           [ cumcount ; function ]
#>           [ cummax ; function ]
#>           [ cummin ; function ]
#>           [ cumprod ; function ]
#>           [ cumsum ; function ]
#>           [ diff ; function ]
#>           [ div ; function ]
#>           [ dot ; function ]
#>           [ drop_nans ; function ]
#>           [ drop_nulls ; function ]
#>           [ eq ; function ]
#>           [ ewm_mean ; function ]
#>           [ ewm_std ; function ]
#>           [ ewm_var ; function ]
#>           [ exclude ; function ]
#>           [ exp ; function ]
#>           [ explode ; function ]
#>           [ extend_constant ; function ]
#>           [ extend_expr ; function ]
#>           [ fill_nan ; function ]
#>           [ fill_null ; function ]
#>           [ filter ; function ]
#>           [ first ; function ]
#>           [ flatten ; function ]
#>           [ floor ; function ]
#>           [ forward_fill ; function ]
#>           [ gt ; function ]
#>           [ gt_eq ; function ]
#>           [ hash ; function ]
#>           [ head ; function ]
#>           [ inspect ; function ]
#>           [ interpolate ; function ]
#>           [ is_between ; function ]
#>           [ is_duplicated ; function ]
#>           [ is_finite ; function ]
#>           [ is_first ; function ]
#>           [ is_in ; function ]
#>           [ is_infinite ; function ]
#>           [ is_nan ; function ]
#>           [ is_not ; function ]
#>           [ is_not_nan ; function ]
#>           [ is_not_null ; function ]
#>           [ is_null ; function ]
#>           [ is_unique ; function ]
#>           [ keep_name ; function ]
#>           [ kurtosis ; function ]
#>           [ last ; function ]
#>           [ len ; function ]
#>           [ limit ; function ]
#>           [ list ; function ]
#>           [ lit ; function ]
#>           [ log ; function ]
#>           [ log10 ; function ]
#>           [ lower_bound ; function ]
#>           [ lt ; function ]
#>           [ lt_eq ; function ]
#>           [ map ; function ]
#>           [ map_alias ; function ]
#>           [ max ; function ]
#>           [ mean ; function ]
#>           [ median ; function ]
#>           [ min ; function ]
#>           [ mode ; function ]
#>           [ mul ; function ]
#>           [ n_unique ; function ]
#>           [ nan_max ; function ]
#>           [ nan_min ; function ]
#>           [ neq ; function ]
#>           [ null_count ; function ]
#>           [ or ; function ]
#>           [ over ; function ]
#>           [ pct_change ; function ]
#>           [ pow ; function ]
#>           [ prefix ; function ]
#>           [ print ; function ]
#>           [ product ; function ]
#>           [ quantile ; function ]
#>           [ rank ; function ]
#>           [ rechunk ; function ]
#>           [ reinterpret ; function ]
#>           [ rep ; function ]
#>           [ rep_extend ; function ]
#>           [ repeat_by ; function ]
#>           [ reshape ; function ]
#>           [ reverse ; function ]
#>           [ rolling_max ; function ]
#>           [ rolling_mean ; function ]
#>           [ rolling_median ; function ]
#>           [ rolling_min ; function ]
#>           [ rolling_quantile ; function ]
#>           [ rolling_skew ; function ]
#>           [ rolling_std ; function ]
#>           [ rolling_sum ; function ]
#>           [ rolling_var ; function ]
#>           [ round ; function ]
#>           [ rpow ; function ]
#>           [ sample ; function ]
#>           [ search_sorted ; function ]
#>           [ shift ; function ]
#>           [ shift_and_fill ; function ]
#>           [ shuffle ; function ]
#>           [ sign ; function ]
#>           [ sin ; function ]
#>           [ sinh ; function ]
#>           [ skew ; function ]
#>           [ slice ; function ]
#>           [ sort ; function ]
#>           [ sort_by ; function ]
#>           [ sqrt ; function ]
#>           [ std ; function ]
#>           [ sub ; function ]
#>           [ suffix ; function ]
#>           [ sum ; function ]
#>           [ tail ; function ]
#>           [ take ; function ]
#>           [ take_every ; function ]
#>           [ tan ; function ]
#>           [ tanh ; function ]
#>           [ to_physical ; function ]
#>           [ to_r ; function ]
#>           [ top_k ; function ]
#>           [ unique ; function ]
#>           [ upper_bound ; function ]
#>           [ value_counts ; function ]
#>           [ var ; function ]
#>           [ where ; function ]
#>           [ xor ; function ]
#> 
#> 
#>        LazyFrame ( environment ):
#>           [ collect ; function ]
#>           [ describe_optimized_plan ; function ]
#>           [ describe_plan ; function ]
#>           [ filter ; function ]
#>           [ groupby ; function ]
#>           [ join ; function ]
#>           [ limit ; function ]
#>           [ print ; function ]
#>           [ select ; function ]
#>           [ with_column ; function ]
#>           [ with_columns ; function ]
#> 
#> 
#>        LazyGroupBy ( environment ):
#>           [ agg ; function ]
#>           [ apply ; function ]
#>           [ head ; function ]
#>           [ print ; function ]
#>           [ tail ; function ]
#> 
#> 
#>        Series ( environment ):
#>           [ abs ; function ]
#>           [ add ; function ]
#>           [ alias ; function ]
#>           [ all ; function ]
#>           [ any ; function ]
#>           [ append ; function ]
#>           [ apply ; function ]
#>           [ arg_max ; function ]
#>           [ arg_min ; function ]
#>           [ ceil ; function ]
#>           [ chunk_lengths ; function ]
#>           [ clone ; function ]
#>           [ compare ; function ]
#>           [ cumsum ; function ]
#>           [ div ; function ]
#>           [ dtype ; property function ]
#>           [ flags ; property function ]
#>           [ floor ; function ]
#>           [ is_numeric ; function ]
#>           [ is_unique ; function ]
#>           [ len ; function ]
#>           [ max ; function ]
#>           [ min ; function ]
#>           [ mul ; function ]
#>           [ name ; property function ]
#>           [ print ; function ]
#>           [ rem ; function ]
#>           [ rename ; function ]
#>           [ rep ; function ]
#>           [ series_equal ; function ]
#>           [ shape ; property function ]
#>           [ sort ; function ]
#>           [ std ; function ]
#>           [ sub ; function ]
#>           [ sum ; function ]
#>           [ to_frame ; function ]
#>           [ to_r ; function ]
#>           [ to_r_list ; function ]
#>           [ to_r_vector ; function ]
#>           [ value_counts ; function ]
#>           [ var ; function ]
#>