rpolars options

get, set, reset rpolars options

Arguments

strictly_immutable

bool, default = TRUE, keep rpolars strictly immutable. Polars/arrow is in general pro "immutable objects". However pypolars API has some minor exceptions. All settable property elements of classes are mutable. Why?, I guess python just do not have strong stance on immutability. R strongly suggests immutable objects, so why not make polars strictly immutable where little performance costs? However, if to mimic pypolars as much as possible, set this to FALSE.

named_exprs

bool, default = FALSE, allow named exprs in e.g. select, with_columns, groupby, join. a named expresion will be extended with $alias(name) wildcards or expression producing multiple are problematic due to name collision the related option in py-polars is currently called 'pl.Config.with_columns_kwargs' and only allow named exprs in with_columns (or potentially any method derived there of)

no_messages

bool, default = FALSE, turn of messages

Value

current settings as list

current settings as list

list named by options of requirement function input must satisfy

Details

who likes rpolars package messages? use this option to turn them off.

modifing list takes no effect, pass it to pl$set_rpolars_options get/set/resest interact with internal env rpolars:::rpolars_optenv

setting an options may be rejected if not passing opt_requirements

Examples

#rename columns by naming expression, experimental requires option named_exprs = TRUE
pl$set_rpolars_options(named_exprs = TRUE)
pl$DataFrame(iris)$with_columns(
  pl$col("Sepal.Length")$abs(), #not named expr will keep name "Sepal.Length"
  SW_add_2 = (pl$col("Sepal.Width")+2)
)
#> polars DataFrame: shape: (150, 6)
#> ┌──────────────┬─────────────┬──────────────┬─────────────┬───────────┬──────────┐
#> │ Sepal.Length ┆ Sepal.Width ┆ Petal.Length ┆ Petal.Width ┆ Species   ┆ SW_add_2 │
#> │ ---          ┆ ---         ┆ ---          ┆ ---         ┆ ---       ┆ ---      │
#> │ f64          ┆ f64         ┆ f64          ┆ f64         ┆ cat       ┆ f64      │
#> ╞══════════════╪═════════════╪══════════════╪═════════════╪═══════════╪══════════╡
#> │ 5.1          ┆ 3.5         ┆ 1.4          ┆ 0.2         ┆ setosa    ┆ 5.5      │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
#> │ 4.9          ┆ 3.0         ┆ 1.4          ┆ 0.2         ┆ setosa    ┆ 5.0      │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
#> │ 4.7          ┆ 3.2         ┆ 1.3          ┆ 0.2         ┆ setosa    ┆ 5.2      │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
#> │ 4.6          ┆ 3.1         ┆ 1.5          ┆ 0.2         ┆ setosa    ┆ 5.1      │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
#> │ ...          ┆ ...         ┆ ...          ┆ ...         ┆ ...       ┆ ...      │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
#> │ 6.3          ┆ 2.5         ┆ 5.0          ┆ 1.9         ┆ virginica ┆ 4.5      │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
#> │ 6.5          ┆ 3.0         ┆ 5.2          ┆ 2.0         ┆ virginica ┆ 5.0      │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
#> │ 6.2          ┆ 3.4         ┆ 5.4          ┆ 2.3         ┆ virginica ┆ 5.4      │
#> ├╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌╌┼╌╌╌╌╌╌╌╌╌╌┤
#> │ 5.9          ┆ 3.0         ┆ 5.1          ┆ 1.8         ┆ virginica ┆ 5.0      │
#> └──────────────┴─────────────┴──────────────┴─────────────┴───────────┴──────────┘
 pl$get_rpolars_options()
#> $strictly_immutable
#> [1] FALSE
#> 
#> $named_exprs
#> [1] TRUE
#> 
#> $no_messages
#> [1] FALSE
#> 
pl$set_rpolars_options(strictly_immutable = FALSE)
pl$get_rpolars_options()
#> $strictly_immutable
#> [1] FALSE
#> 
#> $named_exprs
#> [1] TRUE
#> 
#> $no_messages
#> [1] FALSE
#> 

#setting strictly_immutable = 42 will be rejected as
tryCatch(
  pl$set_rpolars_options(strictly_immutable = 42),
  error= function(e) print(e)
)
#> <simpleError: option: strictly_immutable  must satisfy requirement named is_bool with function
#>  function (x) {
#>     is.logical(x) && length(x)==1 && !is.na(x)
#>   }
#> <environment: namespace:rpolars>>

#reset options like this
pl$reset_rpolars_options()
#use get_rpolars_opt_requirements() to requirements
pl$get_rpolars_opt_requirements()
#> $strictly_immutable
#> $strictly_immutable$is_bool
#> function (x) {
#>     is.logical(x) && length(x)==1 && !is.na(x)
#>   }
#> <environment: namespace:rpolars>
#> 
#> 
#> $named_exprs
#> $named_exprs$is_bool
#> function (x) {
#>     is.logical(x) && length(x)==1 && !is.na(x)
#>   }
#> <environment: namespace:rpolars>
#> 
#> 
#> $no_messages
#> $no_messages$is_bool
#> function (x) {
#>     is.logical(x) && length(x)==1 && !is.na(x)
#>   }
#> <environment: namespace:rpolars>
#> 
#>