Inner workings of the DataFrame-class

The DataFrame-class is simply two environments of respectively the public and private methods/function calls to the rpolars rust side. The instanciated DataFrame-object is an externalptr to a lowlevel rust polars DataFrame object. The pointer address is the only statefullness of the DataFrame object on the R side. Any other state resides on the rust side. The S3 method .DollarNames.DataFrame exposes all public $foobar()-methods which are callable onto the object. Most methods return another DataFrame-class instance or similar which allows for method chaining. This class system in lack of a better name could be called "environment classes" and is the same class system extendr provides, except here there is both a public and private set of methods. For implementation reasons, the private methods are external and must be called from rpolars:::.pr.$DataFrame$methodname(), also all private methods must take any self as an argument, thus they are pure functions. Having the private methods as pure functions solved/simplified self-referential complications.

Details

Check out the source code in R/dataframe_frame.R how public methods are derived from private methods. Check out extendr-wrappers.R to see the extendr-auto-generated methods. These are moved to .pr and converted into pure external functions in after-wrappers.R. In zzz.R (named zzz to be last file sourced) the extendr-methods are removed and replaced by any function prefixed DataFrame_.

Examples

#see all exported methods
ls(rpolars:::DataFrame)
#>  [1] "as_data_frame" "clone"         "columns"       "dtypes"       
#>  [5] "filter"        "get_column"    "get_columns"   "groupby"      
#>  [9] "height"        "join"          "lazy"          "limit"        
#> [13] "print"         "schema"        "select"        "shape"        
#> [17] "to_data_frame" "to_list"       "to_series"     "to_struct"    
#> [21] "unnest"        "width"         "with_column"   "with_columns" 

#see all private methods (not intended for regular use)
ls(rpolars:::.pr$DataFrame)
#>  [1] "by_agg"                 "clone_see_me_macro"     "columns"               
#>  [4] "dtypes"                 "get_column"             "get_columns"           
#>  [7] "lazy"                   "new"                    "new_with_capacity"     
#> [10] "print"                  "schema"                 "select"                
#> [13] "select_at_idx"          "set_column_from_robj"   "set_column_from_series"
#> [16] "set_column_names_mut"   "shape"                  "to_list"               
#> [19] "to_list_tag_structs"    "to_struct"              "unnest"                

#make an object
df = pl$DataFrame(iris)

#use a public method/property
df$shape
#> [1] 150   5
df2 = df
#use a private method, which has mutability
result = rpolars:::.pr$DataFrame$set_column_from_robj(df,150:1,"some_ints")

#column exists in both dataframes-objects now, as they are just pointers to the same object
# there are no public methods with mutability
df$columns
#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"     
#> [6] "some_ints"   
df2$columns
#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"     
#> [6] "some_ints"   

# set_column_from_robj-method is fallible and returned a result which could be ok or an err.
# No public method or function will ever return a result.
# The `result` is very close to the same as output from functions decorated with purrr::safely.
# To use results on R side, these must be unwrapped first such that
# potentially errors can be thrown. unwrap(result) is a way to
# bridge rust not throwing errors with R. Extendr default behaviour is to use panic!(s) which
# would case some unneccesary confusing and  some very verbose error messages on the inner workings of rust.
unwrap(result) #in this case no error, just a NULL because this mutable method do not return any ok-value
#> NULL

#try unwrapping an error from polars due to unmatching column lengths
err_result = rpolars:::.pr$DataFrame$set_column_from_robj(df,1:10000,"wrong_length")
tryCatch(unwrap(err_result,call=NULL),error=\(e) cat(as.character(e)))
#> Error: Lengths don't match: Could not add column. The Series length 10000 differs from the DataFrame height: 150 
#> when calling:
#>  NULL